slippy2019 drivetrain coding
Apparently i forgot to write any code for the drivetrain, so I added in what needed to be added and pushed it to the robot. The robot would be complete if it wasnt for some wierd happenings on the left intake motor, which refuses to spin, but diddnt earlier. More thursday